Extract ReflectionCard component
